requirements.txt
streamlit>=1.30
pillow>=10.0
numpy>=1.24
pandas>=2.0
torch>=2.1
torchvision>=0.16